Functionality tweak
It is currently a mess but I still wanted you to get a look at it. Again, I'm playing with the search functionality. All together I think I can get the header tabs done in the next few days. What I am coding in python, is a proof of concept app that will be implemented as a feature of this entire app. That will house information about inventory, location, and other subcategories of products. I think it will be fairly easy to integrate that within the app and a lot of what is on here is testing. I may not even continue with the search in java. Progress nonetheless